S11P12A609-414 fix: 일정 삭제시 연관된 게임기록 중 음성파일이 있는 경우에만 S3버킷 객체 삭제

See merge request s11-webmobile2-sub2/S11P12A609!294

+++ imnotdurnk_backend/src/main/java/com/imnotdurnk/domain/calendar/service/CalendarServiceImpl.java
@@ -312,7 +312,9 @@ public class CalendarServiceImpl implements CalendarService {
         Optional<List<GameLogEntity>> gameLogEntities = gameLogRepository.findByCalendarEntity_Id(planId);
         for(GameLogEntity gameLogEntity : gameLogEntities.get()) {
             VoiceEntity voice = voiceRepository.findByLogId(gameLogEntity.getId());
-            s3FileUploadService.deleteFile(voice.getFileName());
+            if(voice != null) {
+                s3FileUploadService.deleteFile(voice.getFileName());
+            }
         }
 
         // 일정 삭제, 해당하는 게임기록과 음성기록 모두 삭제(cascade)
